Enhancement: Extract IsIdenticalTest
