+++ tests/Constraint/IsIdenticalTest.php
@@ -1,0 +1,102 @@
+<?php
+/*
+ * This file is part of sebastian/phpunit-framework-constraint.
+ *
+ * (c) Sebastian Bergmann <sebastian@phpunit.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PHPUnit\Framework\Constraint;
+
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestFailure;
+
+class IsIdenticalTest extends TestCase
+{
+    public function testConstraintIsIdentical()
+    {
+        $a = new \stdClass;
+        $b = new \stdClass;
+
+        $constraint = new IsIdentical($a);
+
+        $this->assertFalse($constraint->evaluate($b, '', true));
+        $this->assertTrue($constraint->evaluate($a, '', true));
+        $this->assertEquals('is identical to an object of class "stdClass"', $constraint->toString());
+        $this->assertCount(1, $constraint);
+
+        try {
+            $constraint->evaluate($b);
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+Failed asserting that two variables reference the same object.
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    public function testConstraintIsIdentical2()
+    {
+        $a = new \stdClass;
+        $b = new \stdClass;
+
+        $constraint = new IsIdentical($a);
+
+        try {
+            $constraint->evaluate($b, 'custom message');
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+custom message
+Failed asserting that two variables reference the same object.
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    public function testConstraintIsIdentical3()
+    {
+        $constraint = new IsIdentical('a');
+
+        try {
+            $constraint->evaluate('b', 'custom message');
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+custom message
+Failed asserting that two strings are identical.
+--- Expected
++++ Actual
+@@ @@
+-'a'
++'b'
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+}
