+++ tests/Framework/ConstraintTest.php
@@ -783,90 +783,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsIdentical()
-    {
-        $a = new \stdClass;
-        $b = new \stdClass;
-
-        $constraint = Assert::identicalTo($a);
-
-        $this->assertFalse($constraint->evaluate($b, '', true));
-        $this->assertTrue($constraint->evaluate($a, '', true));
-        $this->assertEquals('is identical to an object of class "stdClass"', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate($b);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that two variables reference the same object.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintIsIdentical2()
-    {
-        $a = new \stdClass;
-        $b = new \stdClass;
-
-        $constraint = Assert::identicalTo($a);
-
-        try {
-            $constraint->evaluate($b, 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that two variables reference the same object.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintIsIdentical3()
-    {
-        $constraint = Assert::identicalTo('a');
-
-        try {
-            $constraint->evaluate('b', 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that two strings are identical.
---- Expected
-+++ Actual
-@@ @@
--'a'
-+'b'
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintIsNotIdentical()
     {
         $a = new \stdClass;
